refactor: move TableWidget to separate file
