+++ jupyterlab_wipp/src/types.ts
@@ -1,0 +1,5 @@
+// Generic interface for any type of WIPP Collections
+export interface IGenericCollection {
+    id: string,
+    [key: string]: any
+}
