+++ jupyterlab_wipp/tsconfig.json
@@ -20,5 +20,5 @@
     "target": "es2017",
     "types": []
   },
-  "include": ["src/*"]
+  "include": ["src/**/*"]
 }
